refactor: hash module refactor
