+++ src/comments/services/comments.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { CommentsService } from '../comments.service';
 import { UserService } from 'src/user/services/user/user.service';
 import { PrismaService } from 'src/database/prisma.service';
-import { HashService } from 'src/hash/service/hash/hash.service';
+import { HashService } from 'src/hash/service/hash.service';
 import { ProjectsService } from 'src/projects/services/projects/projects.service';
 import { LoggerService } from 'src/logger/logger.service';
 import { Comment } from 'src/comments/types/comments';
