+++ src/hash/hash.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { HashService } from './service/hash/hash.service';
+import { HashService } from './service/hash.service';
 
 @Module({
   providers: [HashService],
