+++ src/hash/service/hash.service.spec.ts
