+++ src/hash/service/hash.service.ts
